backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
scipy==1.11.4
torch==2.1.1
joblib==1.3.2
cachetools==5.3.2
sqlalchemy==2.0.23
pytest==7.4.3
python-dotenv==1.0.0
python-multipart==0.0.6
websockets==11.0.3
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
umap-learn==0.5.4
sentry-sdk[fastapi]==1.38.0
jinja2==3.1.2
matplotlib==3.7.3
seaborn==0.12.2
loguru==0.7.2
prometheus-client==0.19.0

# PHASE 9: Security & Rate Limiting
aiohttp==3.9.1
starlette==0.27.0
